+++ CHANGELOG.md
@@ -15,7 +15,7 @@ All notable changes to the package will be documented here.
 
 ## [4.29.10] - 2026-03-27
 
-- **Added** new *color utlities* `applyOpacityToHex` and `percentToHex` for working with *hex colors* and *opacity*.
+- **Added** new *color utilities* `applyOpacityToHex` and `percentToHex` for working with *hex colors* and *opacity*.
 - **Added** new *date utility* `formatDateRelative` to format a date as a relative time string (e.g., `"5m ago"`, `"2h from now"`, etc.) with an optional *custom format* for dates older than a week.
 
 ## [4.29.1] - 2026-03-27